Refactor and move error-tests to 'short'.
Reuse the same cluster for all three tests, which makes them fast enough
to reclassify as 'short' tests.
